Simplify node metric scraping
