+++ pkg/controllers/metrics/pod/controller.go
@@ -180,29 +180,17 @@ func (c *Controller) makeLabels(ctx context.Context, pod *v1.Pod) prometheus.Lab
 	metricLabels[podPhase] = string(pod.Status.Phase)
 	node := &v1.Node{}
 	if err := c.kubeClient.Get(ctx, types.NamespacedName{Name: pod.Spec.NodeName}, node); err != nil {
-		metricLabels[podHostZone] = "N/A"
-		metricLabels[podHostArchitecture] = "N/A"
-		metricLabels[podHostCapacityType] = "N/A"
-		metricLabels[podHostInstanceType] = "N/A"
-		if provisionerName, ok := pod.Spec.NodeSelector[v1alpha5.ProvisionerNameLabelKey]; ok {
-			metricLabels[podProvisioner] = provisionerName
-		} else {
-			metricLabels[podProvisioner] = "N/A"
-		}
+		metricLabels[podHostZone] = ""
+		metricLabels[podHostArchitecture] = ""
+		metricLabels[podHostCapacityType] = ""
+		metricLabels[podHostInstanceType] = ""
+		metricLabels[podProvisioner] = pod.Spec.NodeSelector[v1alpha5.ProvisionerNameLabelKey]
 	} else {
 		metricLabels[podHostZone] = node.Labels[v1.LabelTopologyZone]
 		metricLabels[podHostArchitecture] = node.Labels[v1.LabelArchStable]
-		if capacityType, ok := node.Labels[v1alpha5.LabelCapacityType]; !ok {
-			metricLabels[podHostCapacityType] = "N/A"
-		} else {
-			metricLabels[podHostCapacityType] = capacityType
-		}
+		metricLabels[podHostCapacityType] = node.Labels[v1alpha5.LabelCapacityType]
 		metricLabels[podHostInstanceType] = node.Labels[v1.LabelInstanceTypeStable]
-		if provisionerName, ok := node.Labels[v1alpha5.ProvisionerNameLabelKey]; !ok {
-			metricLabels[podProvisioner] = "N/A"
-		} else {
-			metricLabels[podProvisioner] = provisionerName
-		}
+		metricLabels[podProvisioner] = node.Labels[v1alpha5.ProvisionerNameLabelKey]
 	}
 	return metricLabels
 }
