+++ pkg/metrics/metrics.go
@@ -20,7 +20,7 @@ import (
 )
 
 const (
-	nodeSubsystem    = "nodes"
+	NodeSubsystem    = "nodes"
 	machineSubsystem = "machines"
 )
 
@@ -28,7 +28,7 @@ var (
 	NodesCreatedCounter = prometheus.NewCounterVec(
 		prometheus.CounterOpts{
 			Namespace: Namespace,
-			Subsystem: nodeSubsystem,
+			Subsystem: NodeSubsystem,
 			Name:      "created",
 			Help:      "Number of nodes created in total by Karpenter. Labeled by reason the node was created and the owning provisioner.",
 		},
@@ -40,7 +40,7 @@ var (
 	NodesTerminatedCounter = prometheus.NewCounterVec(
 		prometheus.CounterOpts{
 			Namespace: Namespace,
-			Subsystem: nodeSubsystem,
+			Subsystem: NodeSubsystem,
 			Name:      "terminated",
 			Help:      "Number of nodes terminated in total by Karpenter. Labeled by reason the node was terminated and the owning provisioner.",
 		},
